requirements.txt
python-telegram-bot==13.7
matplotlib==3.4.3
pandas==1.3.3
gql==3.0.0a6
requests==2.26.0
python-dotenv==0.19.0